+++ macbook-installer.zsh
@@ -19,3 +19,6 @@ title
 heading "Installing Xcode Command Line Tools ..."
 xcode-select --install
 
+heading "Installing Homebrew package manager ..."
+/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+