+++ macbook-installer.zsh
@@ -22,3 +22,5 @@ xcode-select --install
 heading "Installing Homebrew package manager ..."
 /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
 
+heading "Installing Homebrew Cask ..."
+brew cask